+++ src/main/java/com/project/entity/Appointment.java
@@ -17,8 +17,8 @@ public class Appointment {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Long appointmentId;
-	private Integer patientId;
-	private Integer doctorId;
+	private Long patientId;
+	private Long doctorId;
 	private LocalDate appointmentDate;
 	private LocalTime appointmentTime;
 	private String status;
